packages: add rc
* codemac/packages/rc.scm: rc implementation created.

There are errors on running make check for rc, as the paths /bin/pwd etc
are not rewritten correctly current. Thus this change removes the check
step from the build of rc - this could be bad if the package is ever
used for a development version. For now this is the released version
where running the unit tests is less interesting.

Signed-off-by: Jeff Mickey <j@codemac.net>